📨 modularized downloadJSON
