+++ src/commands/admin/sendmessage.ts
@@ -1,6 +1,6 @@
 import { MessageEmbed, TextChannel } from 'discord.js';
 import { Command, CommandoClient, CommandoMessage } from 'discord.js-commando';
-import { DownloaderHelper } from 'node-downloader-helper';
+import { downloadJSON } from '../../modules/util/download_attachment';
 
 interface Args {
     channel: TextChannel,
@@ -44,38 +44,30 @@ export default class SendMessageCommand extends Command {
         }
 
         if (msg.attachments.size != 0) {
-            const url = msg.attachments.first()?.url;
-            const is_json = msg.attachments.first()?.name?.endsWith('.json');
-            if (!url || !is_json) {
-                msg.react('❌');
-                msg.reply('Couldn\'t process the attachment.');
-                return null;
-            }
-
-            const download = new DownloaderHelper(url, 'tmp', {
-                override: true,
+            downloadJSON(msg).then(async (json) => {
+                if (!json) {
+                    msg.react('❌');
+                    return msg.reply('Couldn\'t download the attachment.');
+                }
+                try {
+                    await channel.send(content, new MessageEmbed(json));
+                    msg.react('✅');
+                }
+                catch (e) {
+                    msg.react('❌');
+                    msg.channel.send('Couldn\'t send your message: `' + e + '`');
+                }
             });
-
-            download.on('end', async () => {
-                const json_path = download.getDownloadPath().replace('\\', '/');
-                import('./../../../' + json_path)
-                    .then(embed => {
-                        msg.react('✅');
-                        channel.send(content, new MessageEmbed(embed));
-                    })
-                    .catch(err => {
-                        console.log(err);
-                        msg.react('❌');
-                        msg.reply('Couldn\'t download the attachment.');
-                    });
-            });
-
-            download.start();
-
         }
         else {
-            msg.react('✅');
-            channel.send(content);
+            try {
+                msg.react('✅');
+                channel.send(content);
+            }
+            catch (e) {
+                msg.react('❌');
+                msg.channel.send('Couldn\'t send your message: `' + e + '`');
+            }
         }
 
         return null;
